refactor: 회원가입 이벤트 관련 코드 리팩토링
- 클래스 및 변수명을 과거시제로 변경: 이벤트 관련 클래스에는 과거시제를 사용하는 것이 일반적이다.
- 이벤트 핸들러를 관련 도메인 패키지로 이동(User -> Friendship) : 공식 계정을 신규 회원의 친구로 등록하는 핸들러이므로 Friendship 도메인의 기능이다
